+++ Events.php
@@ -6,7 +6,9 @@ use humhub\modules\darkMode\widgets\DarkStyle;
 use humhub\modules\darkMode\widgets\SwitchButton;
 use humhub\modules\darkMode\models\Config;
 use humhub\components\ModuleEvent;
+use humhub\modules\ui\menu\MenuLink;
 use Yii;
+use yii\helpers\Url;
 
 class Events
 {
@@ -17,11 +19,32 @@ class Events
     
     public static function onNotificationAddonInit($event)
     {
-        try {
-            $event->sender->addWidget(SwitchButton::class, [], ['sortOrder' => 200]);
-        } catch (\Throwable $e) {
-            Yii::error($e, 'dark-mode');
+        if (Yii::$app->getModule('dark-mode')->settings->get('showButton', true)) {
+            try {
+                $event->sender->addWidget(SwitchButton::class, [], ['sortOrder' => 200]);
+            } catch (\Throwable $e) {
+                Yii::error($e, 'dark-mode');
+            }
         }
+    }
+    
+    public static function onAccountSettingsMenuInit($event)
+	{
+	    $menu = $event->sender;
+        
+	    $menu->addEntry(new MenuLink([
+	        'icon' => 'fa-moon-o',
+	        'label' => Yii::t('DarkModeModule.base', 'Dark Mode'),
+	        'url' => '#',
+	        'htmlOptions' => [
+	            'data-action-click' => 'ui.modal.load',
+	            'data-action-click-url' => Url::toRoute('/dark-mode/user/'),
+	            'data-pjax-prevent' => ''
+	       ],
+	        'sortOrder' => 900,
+        ]));
+
+        return true;
     }
     
     public static function onDesignSettingForm($event)