+++ messages/de/admin.php
@@ -1,10 +1,13 @@
 <?php
 
 return [
+    '<strong>Dark Mode</strong> module configuration' => '<strong>Dark Mode</strong> Modul-Einstellungen',
     'Adds a Dark Mode to HumHub' => 'Fügt HumHub einen dunklen Modus hinzu',
+    'Button' => 'Button',
     'Current base theme: ' => 'Aktuelles Basis-Theme: ',
     'Dark Mode' => 'Dunkler Modus',
     'Dark Theme' => 'Dunkles Theme',
+    'Show Button in Top Bar' => 'Button in oberer Leiste anzeigen',
     'The stylesheet of the selected theme will be used for the dark mode.' => 'Das Stylesheet des ausgewählten Themes wird für den dunklen Modus verwendet.',
-    '<strong>Dark Mode</strong> module configuration' => '',
+    'Users can set their theme preferences also in Account Settings > General.' => 'Nutzer können die Theme-Einstellung auch in den Account-Einstellungen unter "Allgemein" vornehemen.',
 ];