+++ messages/fa-IR/base.php
@@ -3,6 +3,7 @@
 return [
     'Choose "Follow system" to automatically switch between light and dark mode according to your browser or system preferences.' => '',
     'Dark' => '',
+    'Dark Mode' => '',
     'Follow system (Default)' => '',
     'Light' => '',
     'Theme preferences' => '',