+++ models/Config.php
@@ -16,6 +16,8 @@ class Config extends \yii\base\Model
 
     public $theme;
 
+    public $showButton;
+
     public function init()
     {
         parent::init();
@@ -23,6 +25,8 @@ class Config extends \yii\base\Model
         $settings = Yii::$app->getModule('dark-mode')->settings;
 
         $this->theme = $settings->get('theme');
+
+        $this->showButton = $settings->get('showButton', true);
 
         // If no setting was found, get recommended theme or fallback (DarkHumHub) 
         if (empty($this->theme)) {
@@ -37,20 +41,23 @@ class Config extends \yii\base\Model
     {
         return [
             ['theme', 'in', 'range' => array_keys($this->getThemes())],
+            ['showButton', 'boolean']
         ];
     }
 
     public function attributeLabels()
     {
         return [
-            'theme' => Yii::t('DarkModeModule.admin', 'Dark Theme')
+            'theme' => Yii::t('DarkModeModule.admin', 'Dark Theme'),
+            'showButton' => Yii::t('DarkModeModule.admin', 'Show Button in Top Bar')
         ];
     }
 
     public function attributeHints()
     {
         return [
-            'theme' => Yii::t('DarkModeModule.admin', 'The stylesheet of the selected theme will be used for the dark mode.')
+            'theme' => Yii::t('DarkModeModule.admin', 'The stylesheet of the selected theme will be used for the dark mode.'),
+            'showButton' => Yii::t('DarkModeModule.admin', 'Users can set their theme preferences also in Account Settings > General.')
         ];
     }
 
@@ -88,6 +95,7 @@ class Config extends \yii\base\Model
 
         $settings = Yii::$app->getModule('dark-mode')->settings;
         $settings->set('theme', $this->theme);
+        $settings->set('showButton', $this->showButton);
 
         Yii::$app->cache->delete(DarkStyleAsset::PATH_CACHE);
         Yii::$app->cache->delete(DarkStyleAsset::FILENAME_CACHE);